+++ Android_Coldelabs/kotlin-workmanager/app/build.gradle
@@ -20,9 +20,10 @@ apply plugin: 'kotlin-android-extensions'
 apply plugin: 'kotlin-kapt'
 
 android {
-    compileSdkVersion versions.compileSdk
+    compileSdkVersion 30
+    buildToolsVersion "30.0.3"
     defaultConfig {
-        applicationId "com.example.background"
+        applicationId "com.nudennie.background"
         renderscriptTargetApi 11
         renderscriptSupportModeEnabled true
         vectorDrawables {
@@ -40,6 +41,7 @@ android {
         versionCode 1
         versionName "1.0"
     }
+
     buildTypes {
         release {
             minifyEnabled true
@@ -67,4 +69,8 @@ dependencies {
     implementation "androidx.work:work-runtime-ktx:$versions.work"
     implementation "com.github.bumptech.glide:glide:4.12.0"
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$versions.kotlin"
+
+
+    // WorkManager Dependency
+    implementation "androidx.work:work-runtime-ktx:$versions.work"
 }
