+++ Android_Coldelabs/kotlin-workmanager/app/src/main/java/com/nudennie/background/workers/WorkerUtils.kt
@@ -16,7 +16,7 @@
 
 @file:JvmName("WorkerUtils")
 
-package com.example.background.workers
+package com.nudennie.background.workers
 
 import android.app.NotificationChannel
 import android.app.NotificationManager
@@ -32,14 +32,14 @@ import androidx.renderscript.Allocation
 import androidx.renderscript.Element
 import androidx.renderscript.RenderScript
 import androidx.renderscript.ScriptIntrinsicBlur
-import com.example.background.CHANNEL_ID
-import com.example.background.DELAY_TIME_MILLIS
-import com.example.background.NOTIFICATION_ID
-import com.example.background.NOTIFICATION_TITLE
-import com.example.background.OUTPUT_PATH
-import com.example.background.R
-import com.example.background.VERBOSE_NOTIFICATION_CHANNEL_DESCRIPTION
-import com.example.background.VERBOSE_NOTIFICATION_CHANNEL_NAME
+import com.nudennie.background.CHANNEL_ID
+import com.nudennie.background.DELAY_TIME_MILLIS
+import com.nudennie.background.NOTIFICATION_ID
+import com.nudennie.background.NOTIFICATION_TITLE
+import com.nudennie.background.OUTPUT_PATH
+import com.nudennie.background.R
+import com.nudennie.background.VERBOSE_NOTIFICATION_CHANNEL_DESCRIPTION
+import com.nudennie.background.VERBOSE_NOTIFICATION_CHANNEL_NAME
 import java.io.File
 import java.io.FileNotFoundException
 import java.io.FileOutputStream
@@ -95,7 +95,7 @@ fun sleep() {
     try {
         Thread.sleep(DELAY_TIME_MILLIS, 0)
     } catch (e: InterruptedException) {
-        Log.e(TAG, e.message)
+        Log.e(TAG, e.message.toString())
     }
 
 }
