+++ Android_Coldelabs/kotlin-workmanager/build.gradle
@@ -18,18 +18,17 @@
 
 buildscript {
     ext.versions = [:]
-    versions.compileSdk = 29
     versions.minSdk = 16
-    versions.targetSdk = 29
-    versions.kotlin = '1.4.30'
-    versions.work = "2.5.0"
+    versions.targetSdk = 30
+    versions.kotlin = '1.5.31'
+    versions.work = "2.6.0"
 
     repositories {
         google()
-        jcenter()
+        mavenCentral()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:4.1.2'
+        classpath 'com.android.tools.build:gradle:7.0.3'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$versions.kotlin"
     }
 }
@@ -37,7 +36,7 @@ buildscript {
 allprojects {
     repositories {
         google()
-        jcenter()
+        mavenCentral()
     }
 }
