+++ Android_Coldelabs/kotlin-workmanager/settings.gradle
@@ -1,1 +1,2 @@
+rootProject.name = "WorkManager"
 include ':app'
